fix. table doesn't exist
