--- src/main/java/com/study/springboot/domain/category/controller/CategoryApiController.java
@@ -1,9 +1,0 @@
-package com.study.springboot.domain.category.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class CategoryApiController {
-}
