--- src/main/java/com/study/springboot/domain/category/controller/CategoryViewController.java
@@ -1,7 +1,0 @@
-package com.study.springboot.domain.category.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class CategoryViewController {
-}
