--- src/main/java/com/study/springboot/domain/category/repository/CategoryRepository.java
@@ -1,9 +1,0 @@
-package com.study.springboot.domain.category.repository;
-
-import com.study.springboot.domain.category.entity.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
