--- src/main/java/com/study/springboot/domain/category/service/CategoryService.java
@@ -1,11 +1,0 @@
-package com.study.springboot.domain.category.service;
-
-import com.study.springboot.domain.category.repository.CategoryRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CategoryService {
-    private final CategoryRepository categoryRepository;
-}
