--- src/main/java/com/study/springboot/domain/member/controller/MemberApiController.java
@@ -1,12 +1,0 @@
-package com.study.springboot.domain.member.controller;
-
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class MemberApiController {
-
-}
