--- src/main/java/com/study/springboot/domain/member/controller/MemberViewController.java
@@ -1,8 +1,0 @@
-package com.study.springboot.domain.member.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class MemberViewController {
-}
