+++ src/main/java/com/study/springboot/domain/member/repository/MemberRepository.java
@@ -1,9 +1,9 @@
 package com.study.springboot.domain.member.repository;
 
-
-import com.study.springboot.domain.member.entity.Member;
+import com.study.springboot.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
