+++ src/main/java/com/study/springboot/domain/member/service/MemberService.java
@@ -2,11 +2,17 @@ package com.study.springboot.domain.member.service;
 
 
 import com.study.springboot.domain.member.repository.MemberRepository;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class MemberService {
+
     private final MemberRepository memberRepository;
+
+
 }
