--- src/main/java/com/study/springboot/domain/orderItem/controller/OrderItemViewController.java
@@ -1,7 +1,0 @@
-package com.study.springboot.domain.orderItem.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class OrderItemViewController {
-}
