--- src/main/java/com/study/springboot/domain/orderItem/repository/OrderItemRepository.java
@@ -1,9 +1,0 @@
-package com.study.springboot.domain.orderItem.repository;
-
-import com.study.springboot.domain.orderItem.entity.OrderItem;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OrderItemRepository extends JpaRepository<OrderItem, Long> {
-}
