--- src/main/java/com/study/springboot/domain/orderItem/service/OrderItemService.java
@@ -1,11 +1,0 @@
-package com.study.springboot.domain.orderItem.service;
-
-import com.study.springboot.domain.orderItem.repository.OrderItemRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class OrderItemService {
-    private final OrderItemRepository orderItemRepository;
-}
