+++ src/main/java/com/study/springboot/domain/orderSystem/OrderItems.java
@@ -1,0 +1,31 @@
+package com.study.springboot.domain.orderSystem;
+
+
+import com.study.springboot.domain.member.Member;
+import com.study.springboot.domain.product.Product;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Table
+public class OrderItems {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "product_id")
+    private Product product;
+
+    @ManyToOne
+    @JoinColumn(name = "member_id")
+    private Member member;
+
+    @ManyToOne
+    @JoinColumn(name = "order_list_id")
+    private OrderList orderList;
+}
