+++ src/main/java/com/study/springboot/domain/orderSystem/OrderList.java
@@ -1,0 +1,24 @@
+package com.study.springboot.domain.orderSystem;
+
+
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table
+
+public class OrderList {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @OneToMany(mappedBy = "orderList")
+    private List<OrderItems>  orderItems = new ArrayList<>();
+}
