--- src/main/java/com/study/springboot/domain/orders/controller/OrdersApiController.java
@@ -1,9 +1,0 @@
-package com.study.springboot.domain.orders.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class OrdersApiController {
-}
