--- src/main/java/com/study/springboot/domain/orders/controller/OrdersViewController.java
@@ -1,7 +1,0 @@
-package com.study.springboot.domain.orders.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class OrdersViewController {
-}
