--- src/main/java/com/study/springboot/domain/orders/repository/OrdersRepository.java
@@ -1,9 +1,0 @@
-package com.study.springboot.domain.orders.repository;
-
-import com.study.springboot.domain.orders.entity.Orders;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OrdersRepository extends JpaRepository<Orders, Long> {
-}
