--- src/main/java/com/study/springboot/domain/orders/service/OrdersService.java
@@ -1,11 +1,0 @@
-package com.study.springboot.domain.orders.service;
-
-import com.study.springboot.domain.orders.repository.OrdersRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class OrdersService {
-    private final OrdersRepository ordersRepository;
-}
