+++ src/main/java/com/study/springboot/domain/product/Product.java
@@ -1,0 +1,34 @@
+package com.study.springboot.domain.product;
+
+
+import com.study.springboot.enumeration.CategoryProduct;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table
+
+public class Product {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+
+    @Column
+    private String productName;
+
+    @Column
+    private Integer productPrice;
+
+    @Column String productCode;
+
+
+    @Column
+    @Enumerated
+    private CategoryProduct category;
+
+}
