--- src/main/java/com/study/springboot/domain/product/controller/ProductApiController.java
@@ -1,9 +1,0 @@
-package com.study.springboot.domain.product.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class ProductApiController {
-}
