--- src/main/java/com/study/springboot/domain/product/controller/ProductViewController.java
@@ -1,7 +1,0 @@
-package com.study.springboot.domain.product.controller;
-
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class ProductViewController {
-}
