--- src/main/java/com/study/springboot/domain/product/repository/ProductRepository.java
@@ -1,9 +1,0 @@
-package com.study.springboot.domain.product.repository;
-
-import com.study.springboot.domain.product.entity.Product;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ProductRepository extends JpaRepository<Product, Long> {
-}
