--- src/main/java/com/study/springboot/domain/product/service/ProductService.java
@@ -1,11 +1,0 @@
-package com.study.springboot.domain.product.service;
-
-import com.study.springboot.domain.product.repository.ProductRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class ProductService {
-    private final ProductRepository productRepository;
-}
