+++ src/main/java/com/study/springboot/enumeration/CategoryProduct.java
@@ -1,0 +1,5 @@
+package com.study.springboot.enumeration;
+
+public enum CategoryProduct {
+    BURGER_SET, BURGER_SINGLE, HAPPY_MEAL, DRINK, DESSERT
+}
