+++ src/main/java/com/study/springboot/enumeration/UserRole.java
@@ -1,0 +1,11 @@
+package com.study.springboot.enumeration;
+
+public enum UserRole {
+    USER("ROLE_ADMIN"), ADMIN("ROLE_ADMIN");
+
+    private String value;
+
+    UserRole(String value){
+        this.value = value;
+    }
+}
