+++ src/main/resources/application.properties
@@ -8,6 +8,7 @@ spring.jpa.generate-ddl=true
 spring.jpa.show-sql=true
 spring.jpa.database-platform=org.hibernate.dialect.MySQL8Dialect
 
+
 # mysql
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.hikari.jdbc-url=jdbc:mysql://localhost:3306/kioskdb?characterEncoding=UTF-8&serverTimezone=Asia/Seoul
